[558032] Add Diagram Title Block preferences on Project properties
Bug: 558032
