+++ core/plugins/org.polarsys.capella.core.sirius.analysis/plugin.xml
@@ -602,6 +602,12 @@
           id="org.polarsys.capella.core.sirius.analysis.preferences.DiagramsProcessChainPathPreferencePagePreferencePage"
           name="%Diagram_Process_Chain_Path_Page_Name">
     </page>
+    <page
+	      category="org.polarsys.capella.common.ui.preferences.diagram"
+	      class="org.polarsys.capella.core.sirius.analysis.preferences.TitleBlockPreferencePage"
+	      id="org.polarsys.capella.core.platform.sirius.ui.actions.preferences.TitleBlockPage"
+	      name="%Diagram_Title_Block_Page_Name">
+	</page>
  </extension>
  <extension
        point="org.eclipse.core.runtime.preferences">
@@ -619,7 +625,6 @@
             class="org.polarsys.capella.core.commands.preferences.ui.AbstractCapellaCommandsPreferencePage"
             id="org.polarsys.capella.common.ui.preferences.diagram"
             name="%Diagram_Page_Name">
-            
     </page>
     <page
           category="org.polarsys.capella.common.ui.preferences.diagram"
@@ -632,15 +637,19 @@
                 definitionId="org.polarsys.capella.core.preferences.project.property.pages.visibilityExpression">
           </reference>
        </enabledWhen>
-          
-       
     </page>
-        <page
-    category="org.polarsys.capella.common.ui.preferences.diagram"
-      class="org.polarsys.capella.core.sirius.analysis.preferences.TitleBlockPreferencePage"
-      id="org.polarsys.capella.core.platform.sirius.ui.actions.preferences.TitleBlockPage"
-      name="%Diagram_Title_Block_Page_Name">
-    </page>
+    <page
+	      category="org.polarsys.capella.common.ui.preferences.diagram"
+	      class="org.polarsys.capella.core.sirius.analysis.preferences.TitleBlockPreferencePage"
+	      id="org.polarsys.capella.core.platform.sirius.ui.actions.preferences.TitleBlockPage"
+	      name="%Diagram_Title_Block_Page_Name"
+          objectClass="org.eclipse.core.resources.IProject">
+            <enabledWhen>
+          <reference
+                definitionId="org.polarsys.capella.core.preferences.project.property.pages.visibilityExpression">
+          </reference>
+       </enabledWhen>
+	</page>
  </extension>
  <extension id="org.polarsys.capella.core.sirius.analysis.CapellaInterpreter" point="org.eclipse.sirius.common.expressionInterpreter">
     <expressionInterpreterProvider interpreterProviderClass="org.polarsys.capella.core.sirius.analysis.CapellaInterpreter" />
